game-of-thrones/requirements.txt
asgiref==3.5.0
black==22.3.0
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
distlib==0.3.4
Django==3.2.12
django-cors-headers==3.11.0
django-environ==0.8.1
django-filter==21.1
django-rest-framework==0.1.0
django-stubs==1.10.1
django-stubs-ext==0.4.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
filelock==3.6.0
gunicorn==20.1.0
identify==2.4.12
idna==3.3
importlib-metadata==1.7.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
MarkupSafe==2.1.1
mypy==0.942
mypy-extensions==0.4.3
nodeenv==1.6.0
numpy==1.21.5
oauthlib==3.2.0
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.1
pre-commit==2.17.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.2
types-pytz==2021.3.6
types-PyYAML==6.0.5
typing-extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.14.0
whitenoise==6.0.0
zipp==3.7.0
